Modified signup.tsx
